requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.31.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
PyYAML>=6.0
tqdm>=4.65.0
pillow>=9.0.0
opencv-python>=4.7.0
jupyter>=1.0.0
ipykernel>=6.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
# Optional for experiment tracking
wandb>=0.15.0  # optional
# VLM dependencies
ftfy>=6.1.1
regex>=2023.0.0
open_clip_torch>=2.20.0
flamingo-pytorch>=0.0.1
sentencepiece
peft>=0.5.0
PyWavelets>=1.4.1
packaging>=23.0
